optimization

+++ include/lexer.h
@@ -5,7 +5,7 @@
 #include "vector.h"
 namespace sjtu {
     using std::string;
-    long long stringtoint(const string &);
+    int stringtoint(const string &);
     vector<string> decompose(const string &); //command->pieces
     vector<string> split(const string &); // a|b|c -> (a, b, c)
 }
